pyproject.toml
[tool.poetry]
name = "eeet2582-backend"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["Quynh Le"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.0.0"
django-allauth = "0.51.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^69.0.2"
stripe = "^7.10.0"
django-cors-headers = "^4.3.1"
python-docx = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
